requirements.txt
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-braces==1.13.0
django-cors-headers==3.1.1
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
websocket==0.2.1
websocket-client==0.56.0
whitenoise==4.1.4
